refactor: Show or hide menu tabs based on the role admin or manager
