+++ app/login/page.tsx
@@ -10,6 +10,7 @@ import { ChefHat } from "lucide-react";
 import { useLoginMutation } from "@/lib/store/api";
 import Cookie from "js-cookie";
 import { Role } from "@/types/roles";
+import { jwtDecode } from 'jwt-decode';
 
 export default function LoginPage() {
   const [email, setEmail] = useState("");
@@ -25,15 +26,13 @@ export default function LoginPage() {
       const result = await login({ email, password }).unwrap();
 
       if (result.success) {
+        // Decodificar el token
+        const payload: {email: string, names: string, role: Role} = jwtDecode(result.token);
         Cookie.set("masala-admin-token", result.token);
-        Cookie.set("masala-admin-email", result.user.email);
-        Cookie.set("masala-admin-name", `${result.user.names} ${result.user.lastNameMaternal} ${result.user.lastNamePaternal}`);
-        Cookie.set("masala-admin-role", result.user.role);
-        if(result.user.role === Role.ADMIN) {
-          router.push("/dashboard");
-        } else if(result.user.role === Role.MANAGER) {
-          router.push("/manager");
-        }
+        Cookie.set("masala-admin-email", payload.email);
+        Cookie.set("masala-admin-name", payload.names);
+        Cookie.set("masala-admin-role", payload.role);
+        router.push("/dashboard");
         router.refresh();
       } else {
         toast({
