+++ package.json
@@ -52,6 +52,7 @@
     "eslint-config-next": "13.5.1",
     "input-otp": "^1.2.4",
     "js-cookie": "^3.0.5",
+    "jwt-decode": "^4.0.0",
     "lucide-react": "^0.446.0",
     "next": "13.5.1",
     "next-themes": "^0.3.0",
